Fix crash when listen block was removed while it had connected clients on that port. Reported by AnGeLoCaDuTo (#4544).
There were a few flaws in the code: 1) it should close the listener on /rehash,
shouldnt't matter if there are clients or not, 2) then there was a bug where it
would properly close the listener but it would be re-opened by add_listener2.

Also added a "IRCd no longer listening on .." message if you remove a listen block.