The maximum number of clients (MAXCONNECTIONS) no longer defaults to 1024. The new question in ./Config now defaults to 'auto' (both for new installs and for upgrades). You can still specify a manual limit but it is no longer recommended. A MAXCONNECTIONS of 'auto' means - at present - that UnrealIRCd will try to set a limit of 8192. This is quite a bump from the original 1024. On systems where this is not possible we will simply use the highest amount possible, such as 4096 on many systems, or 1024. In fact, we now no longer error when MAXCONNECTIONS is higher than the 'ulimit -n' limit but will adjust ourselves to the limit. Only if the effective limit is below 100 we will print out a fatal error since running in such a scenario is highly discouraged. The reason for this change is that nowadays with drone attacks we may need to be able to handle more concurrent sockets. Also, many Linux distro's have a default setting of unlimited or 4096 nowadays, out of the box.
For people packaging UnrealIRCd (not end-users):
The ./configure --with-fd-setsize=xx option was removed and the
optional(!!) --with-maxconnections=xx option has been added.
We recommend you NOT to pass this option. Not passing it means that
the previously mentioned 'auto' mode will be used, which is likely
best for most users.

Module coders:
Although it is unlikely you accessed the 'MAXCLIENTS' variable,
if you did, it is now called 'maxclients' (lowercase) since it is
adjusted at runtime and no longer a macro.