BIG changes internally that will break all modules and required many code changes in UnrealIRCd itself: 1) Clients are no longer freed directly by exit_client. Most fields    are freed, but 'sptr' itself is not, so you can use IsDead() on it. 2) exit_client now returns void rather than int 3) ALL command functions return void rather than int.    Of course this also affects do_cmd, command overrides, etc.
This is a direct consequence of the removal of 'cptr' earlier, as that
was used to signal certain things that are now no longer possible
(and it raises the question if things were always correctly signaled
in the first place, so may fix some bugs).
It also makes the code more resillient against cases where you forgot
to check if the client was freed. Still, you are encouraged to do an
IsDead(sptr) if you are calling functions that may kill clients,
such as command functions or things that may use spamfilter.

More changes will follow, such as the removal of FLUSH_BUFFER.