Get rid of pseudo ID's. This also means that SASL with anope won't work anymore if you run latest anope 2.0.6. You need the fix from Feb 9, 2019: https://github.com/anope/anope/commit/da6e2730c259d6d6356a0a948e85730ae34663ab (.. which also fixes SASL problems with anope + UnrealIRCd 4 by the way) or just run anope latest git (2.0 branch).
Not sure about atheme... should test this.

Technical details: we used a pseudo ID / sasl cookie until recently,
this has always been planned to be phased out when we got UID's.
I didn't phase it out in U4 (but could have done so) but just did now in U5.
This simplifies everything as now you can just refer from the services
side to the user with the UID/SID. This also makes it so services can now
target the user in other functions as well, like NOTICE.
(Feel free to request other functions if something isn't working)