Fix trailing space in RPL_NAMREPLY
Currently, UnrealIRCd outputs a trailing space to the trailing parameter
of RPL_NAMEREPLY (when there are users in the channel). However,
according to RFC 1459 and RFC 2812, this is forbidden, since the grammar
for this reply is (quoting RFC 2812):
( "=" / "*" / "@" ) <channel> :[ "@" / "+" ]
 <nick> *( " " [ "@" / "+" ] <nick> )

This commit fixes the issue by adding the joining space only if there is
another, remaining nick to be appended to the buffer.