Fix crash of currently running IRCd when you are running ./Config. Reported by k4be and others. For the crash to occur a few specific things had to happen: 1) The system is missing the argon2 dev library (or it is too old)    causing us to use the UnrealIRCd-shipped argon2 library. 2) You ran ./Config while there is an existing IRCd running 3) Now some argon2 hash is being checked (eg due to an OPER attempt) 4) Crash
A very similar crash happens (to a LOT more people) when you
run './unrealircd restart' to do the actual upgrade. In such
a case, the old IRCd crashed (the one that was actually supposed
to die anyway). The annoying thing was that the crash reporter
would kick in to report such a crash which was actually quite
harmless. This is actually the same crash as described earlier
so should be fixed as well now.
This variant was reported by Shillos and others.