Remove 435 BANNICKCHANGE support
This was only ever of dubious legitimacy and it interacts poorly with
qbans (quiets). For example, a channel which only wants registered users
to speak may set +q ~q:~a:*. If there are a problem with services, or
someone reconnects and auto-joins many channels, they will have to
manually part and rejoin any channels using this approach before they
can change their nick.

This also removes the ~n extban, under a rationale which applies to
both: global settings like the user's nick are beyond the channel
operator's domain of privilege to control of their channel members.

Even if one holds these features to be legitimate, they are trivially
subverted - leave the channel, change your nick, then return - and
therefore are not justified in the face of the problems they introduce.
If someone is using nick changes to subvert a mute of another sort, then
an affected operator should naturally escalate the mute to a ban to
correct the problem.

The global channel mode +N has been left in, as it's only ever used as a
temporary measure against floods and is not considered reasonable to
leave on permanently.