Warn user when undocumented set::ssl::dh / set::tls::dh is present. That option specified a Diffie Hellman parameter file. Since UnrealIRCd 5.0.0 we no longer process this option. This option has never been documented in the wiki docs. We prefer and use ECDHE/EECDH with SSL_OP_SINGLE_ECDH_USE since 2015 to provide Forward Secrecy in SSL/TLS. And indeed, by now in 2020, any properly maintained software uses it and old DH(E) usage has fallen to less than 1%.
What this patch does is remove the unused code (since Dec 2019) and
show a warning if you have a ::dh config directive, so that at least
you are informed that it is unused/ignored. Since it was undocumented
it probably hardly affects anyone, but still, it is proper to inform.