For SJOIN, send the shorter nomode&nopara version when possible. We used to always send the long version: SJOIN ts #channel +sntkl key 999 :xxx From now on we only send that for the first SJOIN for a channel when syncing. For any subsequent SJOINs (so for larger channels or with lots of bans/exempts/invexes) we will use the short version: SJOIN ts #channel :xxx
We now do it that way because the remote side already received
all the modes the first time, so they are redundant in the
subsequent SJOINs for the same channel.
Especially if you have a channel with a large mode string, such as
+lLfH 99 #redirectchan [30j#i10,40m#m10,7c#C15,10n#N15,30k#K10]:15 100:1d
it was previously 1) wasting bandwidth and 2) unnecessary CPU
trying to merge channel modes that were already the same.