Lower libmaxminddb requirement from 1.6.0 to 1.2.0
libmaxminddb 1.6.0 was released on 2021-04-29 and is not yet widely
available in Linux distributions. However, the alternative GeoIP Legacy
C Library will be retired at the end of May 2022 and receives currently
only critical security and bug fixes. Upstream recommends to upgrade to
GeoIP2 databases, which leads to libmaxminddb when using C.

See also:
 - https://github.com/maxmind/geoip-api-c/blob/main/README.md
 - https://repology.org/project/libmaxminddb/versions