When unloading any of the vhoaq channel modes it would cause a crash later on, because unload_extcmode_commit() would call extcmode_para_delslot() even though member modes don't use a parameter slot, and hence it would NULLify a wrong slot, usually for the +H parameter mode. Fun.
We now no longer crash and mass-unset the modes on everyone in the
channel when such a mode is unloaded, just like we do when unloading
any of the other channel modes. It is not done in an efficient way
(one mode per line) but this should be an extremely rare event anyway.

Crash reported by CrazyCat.