When using the "let unrealircd compile cURL" option, now no longer use c-ares, as it complicates things too much. The c-ares build options that we use in UnrealIRCd cause curl not to recongize c-ares, and the other way around is not good either. Also, self-compiled ("unrealircd shipped") c-ares may not be used/required by main unrealircd (thus rm'd) while it is still needed by self-compiled curled. Blehh, what a mess.
Now we simply don't compile curl with c-ares and rely on cURL to enable async
DNS support via system c-ares or via the another way, with the use of threads,
which is standard in curl now for many years and should work on most, if not
all platforms.
If this is somehow problematic for you then install libcurl/libcurl-dev(el)
on your system itself, via your package manager or other means.