WHOWAS: Show IP address and account to IRCOps. Thanks to Noisytoot for https://github.com/unrealircd/unrealircd/pull/227 who suggested displaying account and provided a partial patch, and armyn in https://bugs.unrealircd.org/view.php?id=6153 suggesting IP.
I chose to use the existing RPL_WHOIS* numerics that we also use for
returning WHOIS data. We already use RPL_WHOISSERVER in WHOWAS for
ages and the use of it is mentioned in RFC1459, so seems like that
was the idea right from the beginning of times. The only change I did
was from "is" to "was" in like "was logged in" and "was connecting from"
in the text of the numerics.